% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapsibleTree.R
\name{collapsibleTree.data.frame}
\alias{collapsibleTree.data.frame}
\title{Create Interactive Collapsible Tree Diagrams from a \code{data.frame}}
\usage{
\method{collapsibleTree}{data.frame}(df, hierarchy,
  root = deparse(substitute(df)), inputId = NULL, width = NULL,
  height = NULL, attribute = "leafCount", aggFun = sum,
  fill = "lightsteelblue", fillByLevel = TRUE, linkLength = NULL,
  fontSize = 10, tooltip = FALSE)
}
\arguments{
\item{df}{a data frame from which to construct a nested list}
}
\description{
Create Interactive Collapsible Tree Diagrams from a \code{data.frame}
}
