% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapsibleTree.R
\name{collapsibleTree.Node}
\alias{collapsibleTree.Node}
\title{Create Interactive Collapsible Tree Diagrams from a \code{data.tree}}
\usage{
\method{collapsibleTree}{Node}(df, hierarchy_attribute = "Group",
  root = df$name, inputId = NULL, width = NULL, height = NULL,
  attribute = "leafCount", aggFun = sum, fill = "lightsteelblue",
  fillByLevel = TRUE, linkLength = NULL, fontSize = 10, tooltip = FALSE)
}
\arguments{
\item{df}{a data tree from which to a collapsible tree diagram}
}
\description{
Create Interactive Collapsible Tree Diagrams from a \code{data.tree}
}
