% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapsibleTree.R
\name{collapsibleTree}
\alias{collapsibleTree}
\title{Create Interactive Collapsible Tree Diagrams}
\source{
Christopher Gandrud: \url{http://christophergandrud.github.io/networkD3/}.

d3noob: \url{https://bl.ocks.org/d3noob/43a860bc0024792f8803bba8ca0d5ecd}.
}
\usage{
collapsibleTree(df, hierarchy, root = "Root", width = NULL, height = NULL)
}
\arguments{
\item{df}{a data frame from which to construct a nested list}

\item{hierarchy}{a vector of column names that define the order
and hierarchy of the tree network}

\item{root}{label of the root node}

\item{width}{width in pixels (optional, defaults to automatic sizing)}

\item{height}{height in pixels (optional, defaults to automatic sizing)}
}
\description{
Interactive Reingold–Tilford tree diagram created using D3.js,
where every node can be expanded and collapsed by clicking on it.
}
\examples{
collapsibleTree(warpbreaks,c("wool","tension","breaks"))

}
